Update patch set 1
Patch Set 1:

extract-files.sh: It copies what ends up being the same copy of CE147F02.bin from two different locations off the device via adb into the same directory in the source tree, so the second one overwrites the first. I suppose that if there's not supposed to be two copies of it on the device, then one of them should go away.

In the kernel source for drivers/media/video/ce147.c from the Korean version of this phone, the SHW-M110S, which has a front facing camera (FFC), there is a provision for two different versions of this camera firmware to be present in the phone's filesystem. There's a test for the camera SOC hardware model that determines which one to upload to that SOC when it reflashes it's firmware. During experimentation with the FFC mod-ed Vibrant, it was determined that one of those firmwares worked better than the other in terms of eliminating the low-light banding effect everyone complained of when the FFC mod was first becoming popular.


copy-files-from-nightly.sh: How is anyone supposed to bootstrap a development tree without first flashing a working phone with a potentially non-working nightly unless there's a script to copy the files from a nightly build?

Hmmm. What if I reworked the patch and created a single script that has a switch that determines whether it copies from a nightly build or from a device?

Patch-set: 1