Update patch set 1
Patch Set 1:

copy-files-from-nightly: the files will be in the same spot. Vendor proprietary files don't change much, and so there is no need for this. The style of the file is the same across all phone, we don't want to change it and don't see the need.

Send the patch for the extract-files.sh separately and let me look at it again.

Patch-set: 1