Update patch set 1
Patch Set 1:

It is *not* bullshit. It's working for me, and has been since July 6th. I get 5 to 6 meter accuracy.

I wonder if using the new bootloader has anything to do with it?

I am able to toggle the GPS on and off, and it works after toggling, which is better than it is in the SimplyHoney 3.9 ROM that I pulled it from.

I wonder if the bcm4329.hcd can make a difference, since the bcm4329 may share the same antenna with GPS and cellular? Look in my repository on github in the android_vendor_samsung repository. The bcm4329.hcd is different between CM and SimplyHoney. I'm using the one from SimplyHoney.

Patch-set: 1