Update patch set 1
Patch Set 1:

I agree that it does not get a lock as quickly as it should. It definitely sees satellites right away, and if you are patient, you'll get a lock.

I find that the compass calibration is off at first, and that I must calibrate it using GPS Status and Toolbox. I start that app, then I set the phone on a flat surface, and use it's tool to calibrate the pitch and roll. Then, I use it to calibrate the compass. I find that in calibrating the compass, I need to hold it on edge and rotate it a few times; that is, hold the phone flat in front of you, tip it to the left until it's vertical with the long way pointing away from you, then rotate it in that plane. Also rotate it in all the other axes.

After the pitch and roll and compass are calibrated, use the "manage gps state" button to clear the GPS data, and reload the agps info... then wait for it to get a lock. It can take a few minutes from cold start. Sometimes it's faster to lock if you take a stroll with it; maybe even better going a little faster in a car or bike; I'm not sure.

I am not sure if the choice of modem.bin version affects GPS lock or not. I suppose it could... I don't know for sure if the GPS is on the same chipset as the GSM cellular... or if it's on it's own unit.

I am considering attempting the hardware fix mention in xda-developers, where you open up the phone and slightly bend the contact for the GPS antenna, so that when the phone is put back together, the GPS antenna has a firmer electrical connection.

Patch-set: 1