Update patch set 1
Patch Set 1:

Had raymonddull test the KANG I uploaded to XDA, the resulting TV out was blurry and frozen. Adding the tvouthack executable to the builds lets it refresh, but it's still blurry and black and white.

I'll test more with him or when my TV out cable comes to fix the blurriness and black and white.

Patch-set: 1